Refactor: Move utility functions to utils and inline styles to CSS file
