+++ frontend/src/components/Calendar/WeekEventStyles.css
@@ -1,0 +1,29 @@
+.titleStyle {
+    font-family: Roboto;
+    color: #081821;
+    font-size: 16px;
+    font-weight: 500;
+    line-height: 18.75px;
+    text-align: left;
+    margin-left: 0px;
+    margin-top: 10px;
+    margin-bottom: 0px
+}
+
+.contentWrapperStyle {
+    display: flex;
+    flex-direction: column;
+    margin-left: 28px;
+    margin-top: 0px;
+}
+
+.timingStyle {
+    font-size: 14px;
+    opacity: 0.4;
+    word-break: break-all;
+    white-space: normal;
+    font-weight: 200;
+    line-height: 40px;
+    margin-top: 30px;
+    margin-right: 5px;
+}
